scripts/requirements.txt
python-dotenv==0.7.1
unicodecsv==0.14.1
watson-developer-cloud==1.0.2
